+++ DPUWebProject/src/main/java/com/dpu/controller/CompanyBillingLocationController.java
@@ -37,7 +37,7 @@ public class CompanyBillingLocationController extends MessageProperties {
 	ObjectMapper mapper = new ObjectMapper();
 
 	@RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.GET)
-	public Object getAll(@PathVariable("companyid") int companyId) {
+	public Object getAll(@PathVariable("companyid") Long companyId) {
 		logger.info("[getAll] : Enter : ");
 		String json = null;
 		try {
@@ -75,7 +75,7 @@ public class CompanyBillingLocationController extends MessageProperties {
 	}
 
 	@RequestMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.DELETE)
-	public Object delete(@PathVariable("id") int id) {
+	public Object delete(@PathVariable("id") Long id) {
 		logger.info("[delete] : Enter : Id : "+id);
 		Object obj = null;
 		boolean result = false;
@@ -118,7 +118,7 @@ public class CompanyBillingLocationController extends MessageProperties {
 	}
 
 	@RequestMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.GET)
-	public Object get(@PathVariable("id") int id) {
+	public Object get(@PathVariable("id") Long id) {
 		logger.info("[get] : Enter : Id : "+id);
 		String json = new String();
 		try {
